优化

+++ src/main/java/com/lianjia/model/Recruit.java
@@ -11,8 +11,8 @@ public class Recruit extends Model<Recruit>
 	public boolean validateBelongtoUser(User user)
 	{
 		//该招聘任务所属人
-		int handling_id = getInt("user_id").intValue();
-		int user_id = user.getInt("id").intValue();
+		long handling_id = getLong("user_id").longValue();
+		long user_id = user.getLong("id").longValue();
 		//该发起请求人
 		return handling_id == user_id ? true : false ;
 	}
